[starnix] Deflake starnix_sysrq_integration_test
This test launches a container with a crashing kernel and expects both
the container and the kernel to exit. However, it does so in a
particular order (container then moniker), and the order is actually not
deterministic. If component manager applies the stop action to these
components in the reverse order, the test will hang.

Instead, have the test wait for both monikers together.

Multiply: starnix_sysrq_integration_test

Fixed: 325143035
Change-Id: Ic245f0d917094321382feda9a65d2f45788ceb1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021125
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Gary Bressler <geb@google.com>