[sshd][i18n] Remove ICU data files from sshd-host
We added ICU data so that these files would be available
to command line utilities that devs could run, such as
`fx shell tz-util`. However, on later inspection, it
seems that such binaries have been removed.

This should remove the need for ICU data to reside in
the sshd-host package. It makes a few things simpler
for us, so if it works let's remove it altogether.

Bug: b/330351469
Change-Id: Ie3b771b111543e854d7c0938b21f9e6291fa89f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021142
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>