[dl] LoadModule::Load retrieves file
This CL refactors LoadModule::Load to retrieve the file. This is done
because retrieving the file is a common step for loading the main module
and dependencies; and it's practical to perform both in a single
function.

For now, the dyn phdr and needed count are stored in LoadModule to make
them easy to access during dependency parsing.

Bug: 323419430
Change-Id: I4511986be0f1fae914a7b0495c227fc0b88ff517
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019337
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>