[cpu_profiler] Acknowledge "Stop" Before Writing Data
If a caller is a synchronous fidl client, we need to delay writing data
to the output socket until we've unblocked the caller by Acknowledging
that we've stopped. Otherwise, we can deadlock with the caller waiting
for us to acknowledge and us waiting for the socket to not be full.

With the new kernel assisted sampling, we can now collect enough data in a test run to fill up the socket, running into the deadlock during tests.

Change-Id: Ibc6caa615c39b69d22d57fee774635ee7a041b65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020253
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>