[starnix] Add tests for file creation.
This CL adds a few tests that call open with `O_CREAT|O_TRUNC`, expecting
to get the correct errors on certain filesystems.

If the filesystem is mounted as read-only, it should return `EROFS`.
If the filesystem is writable, but the user has no write access,
`EACCES` should be returned.
If the filesystem is writable, and the user has write access, the call
should succeed.

Bug: 333037315
Change-Id: Ie4152ce5b51463897965141c479d1fde6b9613e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021693
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>