[kernel][sched] Update access to 'non-local' bandwidth parameters
This change removes the capacity, 'timeslice remaining', and expiration
accessors from ThreadBase, as the latter two are dependent on capacity,
and capacity is to become dependent on the state of the entire run
queue. Given that, it makes more sense to move access to this
information to the run queue.

Bug: 328641440
Change-Id: I3cb3bee748028996fbe81d40cd52096b6803470e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017461
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>