[storage] Add conformance tests for opening after rmdir
When we remove a directory, we should not be able to open it at the
path. `openat(fd, ".")`, where fd is references the removed directory,
is expected to pass. We should also be able to call fstat on fd after.

Fix bug in Fxfs where calling open2 with GET_REPRESENTATION attempts to
find the object linked to the removed directory returns early with an
integrity error.

Bug: b/331491118
Test: fx test fxfs-platform-test
Test: fx test fs-directory-tests
Test: fx test fdio-unittest-package
Test: fx test starnix_user_tests
Change-Id: I2af9b94eecf8075fd4b0a3a9ed82df8e4b8093f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015873
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>