[build][python] Reorder union types tried during deserialization
When the serialization module is deserializing a value into a union
type, try to deserialize into non-string scalar and dataclass types
before trying to deserialize into the generic containers, and try the
string and None types last (if part of the union).

Example:

 some_field: Union[str, int, List[int]]
 value: [42]

This would previously have been deserialized into the string "[42]",
and not into a list.

Now it would be deserialized as a list of ints.

Change-Id: If401998b092001a895a9694672e9330fb5288c08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021396
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>