[ffx][target] Improving target show schema
This changes the machine output for `ffx target show` to have a
more specific schema. Previously, the output was effectively a
map, now the keys are moved to be field names so changes in the contents
of the output will be reflected in the schema.

Change-Id: Ifb9c4e5e67cce8d8edbf6111d64ba8334d917a47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021395
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>