[kernel][object_cache] Remove slab control size constant.
Remove the manually computed slab control block size constant. It is a
source of friction when changing the sizes of locks due to internal
structural changes or adding instrumentation.

Add a pair of static accessors that return the sizes of the control block
and slab entries for a given instantiation of the object cache. These
may be used by more advanced allocators to perform slab size selection
based on amortization thresholds or other constraints.

Change-Id: I002b40249be7026f9d600e5868862a471a9c049a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023621
Commit-Queue: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: John Grossman <johngro@google.com>