[sdio] Fix FIDL to Banjo conversion of buffer region structs
Access the VMO ID when the buffer is a VMO ID type, and the VMO handle
when it is a VMO handle type. Without this fix the SDIO driver will try
to use the wrong type and trigger an assert.

Bug: 333726427
Test: CQ
Change-Id: Ibd263e52553b64a1cac2766a4017354a4dabed37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024292
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>