[ffx][trace] Remove from being built-in
Removes ffx-trace from being built in to the main ffx binary, allowing it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes when `ffx trace` is invoked, but makes it so that both the tool and ffx can build faster and with fewer resources making it easier to iterate on them.

Bug: b/329326263
Change-Id: Icc4c8666f8f8ef90f5557d6ca0754aba3a9c8063
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003814
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>