[kernel][arm64][periphmap] try to align mapppings to maximize large pages
Try to naievly align the mappings of the periphmap ranges to try to
utilize large pages. The algorithm is fairly simple so only really work
if the physical address is largely a power of two, but does a better job
than now, where it seems that everything will be misaligned in practice.

Change-Id: I79b726319d7a5c0490d5ebbfa8b693046cf81276
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024927
Reviewed-by: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>