[fuchsia.net.routes] Rename SetProvider -> RouteTable (1/3)
We want to rename the protocol, but we need to do a soft transition
because of out of tree usage. This is the first CL that duplicates
the protocol with the new name.

Bug: 333900329
Change-Id: Iea97943e2bab89cbf70bdc61d843fae697cd8b3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024929
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>