[fidl] Use infallible Endpoints<>::Create
This change migrates callers over to the infallible variant of
endpoint creation.  This reduces boilerplate
checks for the endpoint creation result.

This CL targets roughly half of the calls to
CreateEndpoints, in the fidl tests and benchmarks which
were able to be correctly converted by a script.

Change-Id: Ic15d9488a803247e2bb3a2ec943f8da098ebf3a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1026015
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Garratt Gallagher <garratt@google.com>