[kernel][arm64] DSB and ISB after mapping a kernel page
In the extremely hard to reproduce case except on certain cpus, insert a
set of barriers after mapping a new page in the kernel. Delay the ISB
until the consistency manager is destructed to try to avoid any more
than are necessary.

The DSB ISHST makes sure the page table entry is visible to any future
translation table walker fetches, and the ISB immediately afterward
makes sure the cpu cannot specualatively execute instructions using the
old translation. This was already done when unmapping pages as a side
effect of the DSB+ISB inserted to finalize the TLBI instruction.

Without it, it's possible (and rarely observable on sufficiently
out-of-order cores) for the cpu to throw a translation fault exception
for instructions that happen after the mapping is added. Presumably this
is because the cpu has speculatively executed instructions before the
mapping exists, performs a translation with no page present, latches a
translation fault but holds onto it until the instruction is retired
after the mapping comes into existence. ARM cores are not apparently
required to re-translate the memory fault before throwing a data abort
to work around this, as x86 cores are known to do.

The ARM spec is not terribly clear about the need for the ISB but it
mentions in the section about DSB instructions that without the presence
of FEAT_ETS2 a limited set of things can happen while a DSB is being
executed, and one of them is translations of virtual addresses of loads
and stores of instructions after the DSB.

FEAT_ETS2 seems to tighten this up, but is not present anything but very
new cores.

Tested by running `ut virtual_alloc` in a loop on the affected cores,
which seem to be a great canary.

Change-Id: I3f641d39c6561c05587fe39bd526227a6d4eaf07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021452
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>