[drivers][audio] Reset signalprocessing completers
When a FIDL client reconnects to the aml-g12-audio-composite
driver after disconnecting, the previously pended completers
are not reset, which generates errors such as "WatchTopology
was re-called while the previous call was still pending".

This CL chooses `SignalProcessingConnect` as an appropriate
place (there are a few) to reset these completers.

Bug: 321292607

Test: aml-g12-composite-test

Change-Id: I6f387d17ff183a784241dcadbc1b4f55026502ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025904
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>