[netstack] Don't add IPv4 limited broadcast route
We used to install a route to 255.255.255.255 on every interface as a
hack to allow the DHCP client to send broadcast packets as part of
initial configuration. This is unnecessary due to the client now using
packet sockets (which bypass the routing table entirely), and causes
broadcasts to get routed incorrectly (they should be routed using the
default route instead).

This makes this change on both netstack2 and netstack3.

Also adds an integration test that checks that broadcasts can be
sent / received correctly, which would fail without this change.

Bug: 290794567
Multiply: ipv4-integration
Change-Id: I1c4f3c9fe2aa74b0e0eb8566c486e8addf837c1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020092
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>