[debugdata] Support profile merging with version 9
There is an LLVM upstream change that changes the raw profile format:
https://github.com/llvm/llvm-project/commit/16e74fd48988ac95551d0f64e1b36f78a82a89a2

We have an issue in Rust on-target profile merging when we try to roll
Clang to a revision that uses the new profile format because debugdata
uses InstrProfData.inc that is shipped by the Clang toolchain. This patch adds support of merging profiles in the old layout that uses version 9 to unblock the Clang roll.

Bug: 333945525
Change-Id: Ie3115e77ca1cb3fa77be3e659ccb0a1efb4e3289
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025666
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gulfem Savrun Yeniceri <gulfem@google.com>